# CMA-Backend/apps/api-gateway/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IUser": {
				"description": "User Model — TypeScript Interface & Response DTO\nSchema mới: UUID, username, date_of_birth, is_active",
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"date_of_birth": {
						"type": "string",
						"format": "date-time"
					},
					"avatar_url": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"student",
							"teacher",
							"parent",
							"admin"
						]
					},
					"is_active": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"email",
					"role",
					"is_active",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserDTO": {
				"description": "DTO for creating a new user",
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"student",
							"teacher",
							"parent",
							"admin"
						]
					},
					"avatar_url": {
						"type": "string"
					},
					"date_of_birth": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"username",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserDTO": {
				"description": "DTO for updating a user",
				"properties": {
					"username": {
						"type": "string"
					},
					"avatar_url": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"student",
							"teacher",
							"parent",
							"admin"
						]
					},
					"date_of_birth": {
						"type": "string"
					},
					"is_active": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IContact": {
				"description": "Contact info liên kết với User",
				"properties": {
					"id": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"category_id": {
						"type": "string"
					},
					"contact_value": {
						"type": "string"
					},
					"is_public": {
						"type": "boolean"
					},
					"is_primary": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"user_id",
					"contact_value",
					"is_public",
					"is_primary",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateContactDTO": {
				"description": "DTO for creating a contact",
				"properties": {
					"contact_value": {
						"type": "string"
					},
					"category_id": {
						"type": "string"
					},
					"is_public": {
						"type": "boolean"
					},
					"is_primary": {
						"type": "boolean"
					}
				},
				"required": [
					"contact_value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICategory": {
				"description": "Categories Model — Bảng lookup dùng chung cho toàn hệ thống.\nThay thế hardcoded enums bằng dữ liệu DB-driven.",
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"is_active": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"type",
					"name",
					"is_active"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCategoryDTO": {
				"properties": {
					"type": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"is_active": {
						"type": "boolean"
					}
				},
				"required": [
					"type",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCategoryDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"is_active": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.any_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"FieldErrors": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
			"ValidateError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"fields": {
						"$ref": "#/components/schemas/FieldErrors"
					}
				},
				"required": [
					"name",
					"message",
					"status",
					"fields"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginDTO": {
				"description": "DTO for User Login",
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProfileDTO": {
				"properties": {
					"username": {
						"type": "string"
					},
					"avatar_url": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ChangePasswordDTO": {
				"properties": {
					"currentPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"currentPassword",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"HealthStatus": {
				"description": "Response shape for the health check endpoint",
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"healthy",
							"degraded"
						]
					},
					"uptime": {
						"type": "number",
						"format": "double"
					},
					"timestamp": {
						"type": "string"
					},
					"services": {
						"properties": {
							"cloudinary": {
								"type": "string",
								"enum": [
									"configured",
									"not_configured"
								]
							},
							"database": {
								"type": "string",
								"enum": [
									"connected",
									"disconnected"
								]
							},
							"redis": {
								"type": "string",
								"enum": [
									"connected",
									"disconnected"
								]
							}
						},
						"required": [
							"cloudinary",
							"database",
							"redis"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"uptime",
					"timestamp",
					"services"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "cma-backend-root",
		"version": "1.0.0",
		"description": "Root of CMA Backend Monorepo",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IUser"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all users (Admin only)",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								}
							}
						}
					}
				},
				"description": "Create a new user",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDTO"
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IUser"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Get a single user by ID",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								}
							}
						}
					}
				},
				"description": "Update a user (Admin only)",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Soft-delete a user (Admin only)",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{userId}/contacts": {
			"get": {
				"operationId": "GetUserContacts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IContact"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all contacts for a user",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "AddContact",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IContact"
								}
							}
						}
					}
				},
				"description": "Add a contact to a user",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateContactDTO"
							}
						}
					}
				}
			}
		},
		"/admin/stats": {
			"get": {
				"operationId": "GetDashboardStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get combined dashboard stats (Users, Growth, etc.)",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/admin/health": {
			"get": {
				"operationId": "GetSystemHealth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get system health and service status",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/categories": {
			"get": {
				"operationId": "GetCategories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ICategory"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Lấy tất cả categories (có thể filter theo type)",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"description": "Optional: filter theo type (e.g. 'class_category', 'status')",
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICategory"
								}
							}
						}
					}
				},
				"description": "Tạo category mới (Admin only)",
				"tags": [
					"Categories"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCategoryDTO"
							}
						}
					}
				}
			}
		},
		"/categories/{id}": {
			"get": {
				"operationId": "GetCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICategory"
								}
							}
						}
					}
				},
				"description": "Lấy category theo ID",
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICategory"
								}
							}
						}
					}
				},
				"description": "Cập nhật category (Admin only)",
				"tags": [
					"Categories"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCategoryDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Xóa category (Admin only)",
				"tags": [
					"Categories"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"user": {
											"$ref": "#/components/schemas/Record_string.any_"
										},
										"accessToken": {
											"type": "string"
										}
									},
									"required": [
										"user",
										"accessToken"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								}
							}
						}
					}
				},
				"description": "Xác thực thông tin người dùng và sinh chuỗi Token truy cập.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDTO"
							}
						}
					}
				}
			}
		},
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"user": {
											"$ref": "#/components/schemas/IUser"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"user",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request - Email Exists"
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								}
							}
						}
					}
				},
				"description": "Đăng kí tài khoản mới vào hệ thống.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDTO"
							}
						}
					}
				}
			}
		},
		"/auth/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"description": "Lấy thông tin người dùng hiện tại (dựa trên JWT).",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUser"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"description": "Cập nhật thông tin hồ sơ người dùng hiện tại.",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProfileDTO"
							}
						}
					}
				}
			}
		},
		"/auth/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								}
							}
						}
					}
				},
				"description": "Đổi mật khẩu (bắt buộc đăng nhập).",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordDTO"
							}
						}
					}
				}
			}
		},
		"/health": {
			"get": {
				"operationId": "GetHealth",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HealthStatus"
								}
							}
						}
					}
				},
				"description": "System health check — verifies connectivity of all core dependencies.\nUsed by deployment platforms (Render, Railway, AWS ECS) to determine app status.",
				"tags": [
					"System"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "https://localhost:3000/api"
		}
	]
}